Process by line.
Add the ability to process line by line through a function and to return
arrays to the caller. Adding this complexity is showing the weakness of
the magic function. It made a lot more sense in the context of
**Timezone**, where the arguments are all strings.

With **Cadence** we have a magic function in `async`, but we're not
trying to compress meaning into a single line. With the `$ch` magic
function being invoked multiple times in a single expression, passing
the return values of `$ch` to `$ch`, things look choppy and they do not
look piped.

The best way to express the pipeline is to use method chaining. I've not
used it in a while. I've eschewed it. But, now it looks like the
direction we're going to go.

Closes #9.
Closes #7.
Closes #3.