Implement method chaining.
Implemented a method chaining API. No tests yet, other than the one pipe
test. I'm going to write programs against it for a while, to see how the
API feels, and forgo the testing ceremony for the time being.

Created a `t/echo.js` to test executables, so that tests will work on
Windows. I've not used it in any tests yet.

Implemented the JavaScript string argument parser, so you can type out a
command line in a JavaScript string, instead of building an array. This
will not invoke the system shell, so it will be consistent across
platforms.

Closes #11.
Closes #10.