Add Window instrumentation, maintain own outbox.
Added instrumentation to Window to try to isolate gaps in window
publishing. The `outbox` that is doubling as our reservoir is external
to us and our dear user may decide to call `end` upon it. In this case
our reservoir truncates and we're no longer gathering the messages we
write into a reservoir. We need our own reservoir.