Track conduit series.
Add a series number to the Conduit. The Conduit has become the lowest
layer in the abstraction. When used, it will be the first to receive a
line of streaming JSON off of a socket. If there are missing lines, this
is the place to catch them because we can close the network connection
and reconnect and a utility such as Window can replay items.