Remove throttle.
Can't really imagine how to queue invocations of the responder function
in such a way that it won't block the Conduit since I'm allowing the
user to use their responder function as a pump for the shifter they've
been given. If a there are only a handful of workers, then only a
handful of long running shifters would block the Conduit from starting
new requests, messages would pile up in an incoming queue, memory would
run out.

Now I'm staring a responder for every request. Feeling pretty good about
`Destructible`, its performance.

Need to document the use of the responder function as a `Destructible`
managed work function. It might reduce the number of sub-destructibles
in an application.

Closes #201.