Destroyed only in response to input stream end.
Now only closing when we receive an end of stream. Closing first pushed
a null onto all the responder function inboxes. It then waits
for all the responder functions to end, then for all the outbox queues
to drain, and it then sends a null down the multiplexed queue.

There is no more `destroy`. It is only destroyed when the input shifter
reaches the end of stream. You can destroy the `Conduit` earlier by
truncating the `Shifter` with `Shifter.destroy()`.

Closes #205.